feat: Improve model evaluation with detailed logging and error handling
- Updated `evaluate_model` function in `eval_model.py` to add detailed logging and handle outputs from complex model output objects.
  - Added checks for `logits` attribute and handled cases where model output is a tensor.
  - Added logging statements to debug model output types and shapes.
  - Ensured predictions and labels are logged for better debugging.

- Updated `main.py` to include more comprehensive error handling and logging.
  - Configured logging to display debug-level messages.
  - Added try-except block to handle and log exceptions during model evaluation.
  - Ensured that the model is loaded, moved to the correct device, and validated properly.

- Updated `setup_test_dataloader.py` to correctly map WordNet IDs to class indices using `IMAGENET2012_CLASSES`.
  - Imported `IMAGENET2012_CLASSES` from `classes.py`.
  - Created a reverse mapping from WordNet IDs to class indices.
  - Modified `ImageNetFlatDataset` class to use this mapping for extracting labels.

- Verified that the validation dataset size is correctly logged in `main.py`.
- Confirmed that the evaluation metrics (accuracy, F1 score, recall) are logged and saved correctly.
- Ensured that the confusion matrix is correctly generated and logged.

Overall, these changes enhance the robustness and transparency of the model evaluation process by providing detailed logging, comprehensive error handling, and correct label mapping.
